add migrate-to-w3up bin script instead of old w32023 specific one
